way more changes than should be in a single commit
Almost another complete overhaul of the library's internals. Among the new features are custom types (the interface is still incomplete), a redoing of how help is printed (still incomplete), and the addition of a deinit method to the parser to free any resources that may have been allocated. This commit is very patchy in some places; this will be improved later as more tests are added to identify corner cases that expose bugs