Begin project restructuring + update to post-Writergate
Due to the Zig stdlib's impending std.cli interface, argz is moving
towards a proposed std.cli-like interface, where a struct is provided
as a schema which the parser will convert a set of command-line
arguments into. The main difference is that the parser is based on a
vtable to allow for different (potentially external to argz)
implementations to transform different kinds of data into the expected
argument format.